Move FilteredYAML type out of internal
It turns out that this type is useful outside of this package, which
makes it a good candidate for graduating out of `internal`.
